+++ src/pages/hyeseong/Login/LoginHyesong.scss
@@ -1,4 +1,4 @@
-.container {
+.LoginHyeseong {
   background-color: whitesmoke;
   height: 100vh;
   padding-top: 100px;
