+++ src/pages/hyeseong/Main/UserComment.js
@@ -1,10 +1,10 @@
 import React from 'react';
 
 const UserComment = props => {
-  let { comment } = props;
+  let { userName, comment } = props;
   return (
     <div className="comments">
-      <span className="user-name">wecode</span>
+      <span className="user-name">{userName}</span>
       <span className="commnets-desc">{comment}</span>
     </div>
   );
